semi-graceful handling of error 403
Clicking on the 'Greenlandic Address Reference Register' breadcrumb
leads the user to a rather unhelpful and ugly 403 error that they
can't get out of, short of hitting that browser back button. That
sucks.

Obviously, we'd rather not display that error, but fixing that within
Django is kind of hard. Instead, we make the error much more graceful;
they now get a nice error page that redirects to the main page after
15 seconds.