temporal meta-model: use object UUIDs for references; drop constaints
Previously, we'd use primary keys in references in both the main and
history tables, with foreign key constraints in both tables. Using
them in the history tables was unintentional; what we wanted to do was
use permanent UUIDs, and retain them when the object was deleted.

This requires changing altering ForeignKeys in the registration table,
changing the field and dropping the constraint. We cannot easily
migrate foreign keys from one field to the other, so this necessitates a
re-import of all data.